requirements.txt
fastapi>=0.115.0
uvicorn>=0.34.0
pydantic-settings>=2.7.0
sqlalchemy>=2.0.0
asyncpg>=0.29.0
cryptography>=42.0.0
httpx>=0.27.0
anthropic>=0.40.0
python-jose[cryptography]>=3.3.0
beautifulsoup4>=4.12.0
openai>=1.75.0
langchain>=0.3.0
langchain-openai>=0.2.0
redis[asyncio]>=5.0.0
kiwipiepy>=0.17.0
numpy>=2.0.0
pgvector>=0.3.0
apscheduler>=3.10.0
langchain>=0.3.0
langchain-openai>=0.3.0
langgraph>=0.3.0
langsmith>=0.2.0
